CEA608: support for cc2-cc4, special/extended characters, some formatting
This commit adds several feature enhancements to CEA608 processing.

1. All four CC tracks are now available

There are now four Cea608Stream objects attached to the caption
stream. Each one has some setup code that sets the bytes used
for parsing out track information. Each Cea608Stream receives
every caption packet, but the Cea608Stream will `return` as soon
as possible so that it doesn't spend time processing packets that
target another track.

The active channel is tracked with a global variable, which is set
when a Cea608Stream receives a control code destined for its channel.

2. Special and Extended Characters

Before, the Cea608Stream would consider all special and extended
characters (as defined in CEA608) as unsupported control codes. It will
not properly translate those characters to their unicode equivalent.

3. Formatting

Underline and italics will now be rendered. These are tracked internally
by instance variable `this.formatting_`.

Additionally, PAC rows are now properly interpreted, and text is
inserted into the correct row in the buffer. Pop-on captions should now
render line breaks at the correct spot. Instead of removing _all_ blank
lines, only those at the start or end of the caption are removed.

4. Misc Structure

The code is structured such that support for more features can be added
easily. This will make support for more features easier to add in the
future, if it is desired.

Still unsupported:
- colors
- custom whitespacing on line start/end
- paint-on mode